+++ lib/lightning/application.ex
@@ -8,6 +8,13 @@ defmodule Lightning.Application do
 
   @impl true
   def start(_type, _args) do
+    # mnesia startup
+    :mnesia.stop()
+    :mnesia.create_schema([node()])
+    :mnesia.start()
+    Hammer.Backend.Mnesia.create_mnesia_table(disc_copies: [node()])
+    :mnesia.wait_for_tables([:__hammer_backend_mnesia], 60_000)
+
     # Only add the Sentry backend if a dsn is provided.
     if Application.get_env(:sentry, :included_environments) |> Enum.any?(),
       do: Logger.add_backend(Sentry.LoggerBackend)