+++ priv/repo/migrations/20230113031058_add_failure_alert_to_project_users.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddFailureAlertToProjectUsers do
+  use Ecto.Migration
+
+  def change do
+    alter table(:project_users) do
+      add :failure_alert, :boolean, default: true
+    end
+  end
+end